Factorize out rangeSpec() function from RangeWidget and UpDownWidget to the utils module.
